spsa tune

+++ src/misc.h
@@ -229,7 +229,7 @@ class MultiArray {
 
     constexpr void swap(MultiArray<T, Size, Sizes...>& other) noexcept { data_.swap(other.data_); }
 
-    void load(int16_t*& dataPtr) {
+    void load(int*& dataPtr) {
         for (auto& ele : data_)
         {
             if constexpr (sizeof...(Sizes) == 0)
